Move the api to its own file and at /api/
This way the API at is own dedicated domain in the urls.

Documentation updated accordingly
Unit-tests updated accordingly

Unit-test coverage improved
